Refactoring to mvc done
