+++ models/tourModel.js
@@ -1,0 +1,20 @@
+import mongoose from 'mongoose';
+
+const tourSchema = new mongoose.Schema({
+  name: {
+    type: String,
+    required: [true, 'A Tour Must have a name'],
+    unique: true,
+  },
+  rating: {
+    type: Number,
+    default: 4.5,
+  },
+  price: {
+    type: String,
+    required: [true, 'A Tour Must have a price'],
+  },
+});
+
+const Tour = mongoose.model('Tour', tourSchema);
+export default Tour;
